activate: check that directories exist and are writable
Check that directories exist and are accesible for writing and
entering (wx) before we perform any changes.

Authselect must be run as root so technically the only thing
that can stop us from writing changes is a missing directory.
But we should perform these checks anyway and stop the activation
before we write any changes to the system configuration.