Make sure /etc/dconf/db/distro.d is created
This directory is currently used by gdm to enable smartcard or fingerprint
authentication. It is created by default in Fedora and RHEL systems but
it is not owned by any package. We call make -p in %post phase of the
package installation to ensure that the directory is created even on
systems without gdm (e.g. docker) so we do not fail there and at the
same time the requested authentication works immediately in case gdm
is installed on such systems.

Resolves:
https://github.com/pbrezina/authselect/issues/30