Replace _malloc_zero() function with calloc().
glibc provides calloc() which will allocate memory for an array of
elements of the specified size and memset it all to 0.  You can also use
calloc() on a single element array, which is effectively a
malloc()+memset() operation in one call.

The malloc_zero macro is still present, it's just that it uses calloc()
now instead of the internal _malloc_zero() function.