common: replace the _realloc_array() function with reallocarray() usage.
glibc 2.26 introduced the reallocarray() function from OpenBSD.  This
does exactly what you would think and is a useful companion to calloc().
It is not part of any standard, so to use it you need to include
malloc.h (note this is inconsistent with documentation stating you can
define _GNU_SOURCE and include stdlib.h).

This removes the malloc.c source file as all of the macros now call
glibc memory management functions.

Resolves:
https://github.com/pbrezina/authselect/issues/156