lib: store copy of generated files to state directory
When some changes are done to the currently selected profile template,
authselect detects that the current configuration is no longer valid
because it compares it against the new templates which have changed.

However the configuration itself did not changed an it is still valid.
We need to remember the originally generated files so we can use them
to detect if any configuration has changed but ignore changes in
profile templates themselves.

Resolves:
https://github.com/pbrezina/authselect/issues/51