lib: add user editable nsswitch.conf
Merge profile's nsswitch.conf with a user editable nsswitch.conf file
located at /etc/authselect/user-nsswitch.conf.

Merging is done as:
1) if nss map is specified in profile, use this value
2) else if it is not present in profile, use value from user file

This is done to allow changes in nsswitch.conf without the need to
create a whole new custom profile. Profiles do not have to specify
all nss maps and they should set only those maps that they do care
about.

If the map is set in the profile it is always used thius it does
ensure that the resulting nsswitch.conf looks exactly as requested
by profile with possible user changes in maps that the profile is
not interested in.

Resolves:
https://github.com/pbrezina/authselect/issues/51