lib: ignore changes in comments and spaces when validating files
Those changes are not affecting configuration so they can be simply
ignored and it will allow us to add different comments to different
generated files.

This type of check is done only if original copies of generated files
are missing (which can happen during an upgrade to newer version).
If the files are not missing, the check is done literraly and no
changes are allowed.

Resolves:
https://github.com/pbrezina/authselect/issues/51