lib: use authselect file names for backup
The authselect system dconf files are named 20-authselect
and locks/20-authselect. Because the backup directory is
flat we end up overwriting the first file with the latter
and actually backing up only the locks information.

Now we use dconf-db and dconf-locks names that are used
in /etc/authselect so both files are backed up correctly.

Relates to:
https://github.com/pbrezina/authselect/issues/120