file_copy: use fchmod() and pass only valid mode bits
Currently in `strace` authselect shows strange system calls like:

chmod("/etc/pam.d/system-auth.NenLC7", 0100644)

The mode parameter actually also contains the type bits, because the
`st_mode` member that was previously obtained from `fstat()` is passed
unmodified.

In this commit a bitwise AND is performed to restrict the parameter to
the permission bits only. Furthermore the `chmod` is changed into an
`fchmod` which is more robust and faster in this context.