utils: correctly remove the whole line if `str` does not point to its beginning
The following scenario triggers a bug when the line is not removed completely
and the two lines are merged instead.
```
BEGINNING {if "condition":true|false} END {include if "condition"}
NEXT LINE

->
BEGINNING falseNEXT LINE
```

This is because `match_string` points after the first condition and we
only remove the line to this point. Therefore we need to interate
before `match_string` so we can find the real line start.

Resolves:
https://github.com/authselect/authselect/issues/218