main: Drop an unnecessary `NULL` check before `free()`
From `man free()`:

```
The free() function frees the memory space pointed to by ptr ...  If ptr is NULL, no operation is performed.
```

Obviously there are *tons* of these in the codebase; just doing
this one as a preliminary PR; if accepted I may do some more, or
others can.  Or we could try a coccinelle semantic patch.