Don't try to write to `/var/lib/authselect` if it's read-only
I haven't fully grasped what authselect is trying to do with
the files in `/var/lib/authselect`.  This patch changes things
so that we skip trying to write to them if the path is read-only,
as will occur in `rpm-ostree compose tree` at build time.

I *think* these are backup copies of the files authselect uses to
detect "drift" from the files it wrote at the time of invocation.

If so, I think it's better to store them in `/etc/authselect/_pristine`
or something.  Or, add checksums as xattrs on the files.

In any case, this is now sufficient to get me a booting FCOS
system in concert with https://github.com/coreos/rpm-ostree/pull/3308