lib: drop strict change detection
One of the main authselect features is to refuse to overwrite user
changes to the configuration. To do so, it used to validate files and
directory structure and also files contents.

This however allowed to have a mixed configuration on the system, that
was created by authselect but then modified by user. This is no longer
possible and the behavior has changed in a way that any user changes
made to configuration created by authselect are overwritten without
requiring the *--force* parameter (files contents are no longer
validated).

This will make the configuration consistent - users either use
authselect or not, nothing in between. It also makes authselect work
on ostree server side, where it is not possible to write to /var.