spec: remove unnecessary dependencies
- gawk, findutils and grep are no longer needed in scriptlets
- systemd does not have to be required

PAM and nsswitch.conf from shipped profiles references several
systemd modules. However if those modules are not installed,
they are simply ignored, therefore this dependency can be
omitted.

systemd is usually installed anyway, but it is not necessary to
pull it in on cases when it is not desired, such as in containers.