Added feature with-pam-gnome-keyring.
pam-gnome-keyring starts a gnome-keyring daemon for the user on login.
It is needed start/unlock/lock a gnome-keyring when not running
within a GUI.

This feature is off by default.  Users/Administrators need to enable
this feature if they need it.

Standardize module order amoung all three profiles.

The gnome-keyring module directives are applied in the same
order in the nis, sssd, and winbind profiles.

Also fixed text minor alignment issues.