authselect: automatically run authselect apply-changes on first boot
This installs an authselect.service file that is pulled in by
sysinit.target and automatically runs "authselect apply-changes" on
first boot (i.e. when /etc/ is unpopulated), and applies any available
configuration.

Due to the ConditionFirstBoot= and ConditionPathIsReadWrite=/etc this
has zero effect on systems which do not come up in a first boot mode, or
where /etc/ is immutable.

See: #355