profiles: include "systemd" in shadow + gshadow nsswitch.conf lines
Since systemd v249 (2021) nss-systemd hasn't just been generating
entries in the "passwd" and "group" databases, but in "shadow" and
"gshadow" as well. This is documented in the man page:

https://www.freedesktop.org/software/systemd/man/latest/nss-systemd.html

This is generally useful, since various account validity fields that
userdb/homed maintains are only exported via spwd/sgrp, but most
importantly this makes the "systemd-nspawn --bind-user=" logic work,
which just propagates the hashed user pw into a container.

The patch is pretty simple: in all profiles where passwd/group already
lists "systemd", we also add it to the lines with shadow/spwd, to the
end.