lib: add upgrade parameter to apply-changes
In this case, changes will be applied only when the installed profiles
in /usr have changed. This is done by installing a checksum of all
available profiles, when a profile is selected (or apply-changes is
called) the checksum is copied over to the state directory.

When apply-changes --upgrade is called, the checksum files are compared
and the command updates the configuration only if it is different.

This is a workaround for ostree systems where the rpm scriptlet is not
run and /usr timestamps are set to 0, therefore we can not use that to
achieve similar behavior. The intention is to call apply-changes
automatically only after authselect package is upgraded to avoid excesive
changes to the confiration.

If it was called after each boot, it would be quite possible to apply
broken configuration in case somebody has custom profile selected but
currently in a workin progress state. This solution does not completely
avoid this issue but makes it very unlikely given that authselect is not
updated very often.